migrations/requirements.txt
alembic>=1.13
psycopg2-binary>=2.9
sqlalchemy>=2.0.30
pgvector>=0.3.2
python-dotenv>=1.0
